+++ library/src/main/kotlin/eu/tib/orkg/prototype/configuration/JpaConfiguration.kt
@@ -7,10 +7,12 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 @Configuration
 @EnableJpaRepositories(
     "eu.tib.orkg.prototype.files.adapter.output.jpa.internal",
+    "eu.tib.orkg.prototype.discussions.adapter.output.jpa.internal",
     "eu.tib.orkg.prototype.community.adapter.output.jpa.internal"
 )
 @EntityScan(
     "eu.tib.orkg.prototype.files.adapter.output.jpa.internal",
+    "eu.tib.orkg.prototype.discussions.adapter.output.jpa.internal",
     "eu.tib.orkg.prototype.community.adapter.output.jpa.internal"
 )
 class JpaConfiguration