+++ library/src/main/kotlin/eu/tib/orkg/prototype/discussions/api/compat.kt
@@ -1,0 +1,3 @@
+package eu.tib.orkg.prototype.discussions.api
+
+interface DiscussionUseCases : CreateDiscussionCommentUseCase, RetrieveDiscussionUseCase, DeleteDiscussionCommentUseCase